+++ ChangeLog.md
@@ -13,6 +13,7 @@
  - ENH: I18N - Add new Semantic Type - POSTAL_CODE.POSTAL_CODE_CO - Colombian Postal Code
  - ENH: Add a wrapper task to indicate the version of Gradle required (Issue #11)
  - ENH: Add ability to create Normalized file to capture both words with and without diacritical marks
+ - ENH: Improve detection of DAY.DIGITS and MONTH.DIGITS (implemented in Java as opposed to regexp)
  - ENH: Improve MiddleName/MiddleInitial detection
  - ENH: Improve Street Address detection
  - ENH: Improve Company Name detection
@@ -23,7 +24,7 @@
  - ENH: Add support for pluginOptions
  - ENH: Bump slf4j (2.0.0) and logback-classic (1.4.0)
  - BUG: Should have cleansed and trimmed input before checking it in isValid()
- - BUG: Samples returned by VAT routines were typically not Vlaid
+ - BUG: Samples returned by VAT routines were typically not Valid
 
 ### 9.1.1
  - BUG: Don't output totalNullCount, totalBlankCount, totalMinLength, totalMaxLength if they are unset